refactor(cli): move setupConfiguration to env.go
