+++ cmd/tk/env.go
@@ -3,6 +3,7 @@ package main
 import (
 	"encoding/json"
 	"fmt"
+	"log"
 	"os"
 	"path/filepath"
 	"text/tabwriter"
@@ -15,6 +16,7 @@ import (
 
 	"github.com/grafana/tanka/pkg/jsonnet/jpath"
 	"github.com/grafana/tanka/pkg/kubernetes/client"
+	"github.com/grafana/tanka/pkg/spec"
 	"github.com/grafana/tanka/pkg/spec/v1alpha1"
 	"github.com/grafana/tanka/pkg/tanka"
 	"github.com/grafana/tanka/pkg/term"
@@ -49,6 +51,32 @@ var kubectlContexts = cli.PredictFunc(
 		return c
 	},
 )
+
+func setupConfiguration(baseDir string) *v1alpha1.Environment {
+	_, baseDir, rootDir, err := jpath.Resolve(baseDir)
+	if err != nil {
+		log.Fatalln("Resolving jpath:", err)
+	}
+
+	// name of the environment: relative path from rootDir
+	name, _ := filepath.Rel(rootDir, baseDir)
+
+	config, err := spec.ParseDir(baseDir, name)
+	if err != nil {
+		switch err.(type) {
+		// the config includes deprecated fields
+		case spec.ErrDeprecated:
+			if verbose {
+				fmt.Print(err)
+			}
+		// some other error
+		default:
+			log.Fatalf("Reading spec.json: %s", err)
+		}
+	}
+
+	return config
+}
 
 func envSetCmd() *cli.Command {
 	cmd := &cli.Command{
